api/requirements.txt
fastapi
uvicorn[standard]
sqlmodel
SQLAlchemy
httpx
pydantic
python-dotenv
tenacity
orjson
